mgconfig/requirements.txt
cryptography==45.0.5
python-dotenv==1.1.1
pytz==2024.2
PyYAML==6.0.2
tzlocal==5.2
